+++ test/quantization/test_quantize_eager_ptq.py
@@ -1,24 +1,18 @@
+
 import torch
 import torch.nn as nn
 import torch.nn.quantized as nnq
-import torch.nn.intrinsic as nni
-import torch.nn.intrinsic.quantized as nniq
-import torch.nn.intrinsic.qat as nniqat
 from torch.nn.utils.rnn import PackedSequence
 from torch.quantization import (
     quantize,
     prepare,
     convert,
     prepare_qat,
-    quantize_qat,
-    fuse_modules,
     quantize_dynamic,
     QuantWrapper,
     QuantStub,
     DeQuantStub,
-    QConfig,
     default_qconfig,
-    default_qat_qconfig,
     default_dynamic_qconfig,
     per_channel_dynamic_qconfig,
     float16_dynamic_qconfig,
@@ -26,18 +20,12 @@ from torch.quantization import (
     PerChannelMinMaxObserver,
     QConfigDynamic,
     default_dynamic_quant_observer,
-    FixedQParamsFakeQuantize,
 )
 
 from torch.testing._internal.common_quantization import (
     QuantizationTestCase,
     AnnotatedSingleLayerLinearModel,
     QuantStubModel,
-    ModelForFusion,
-    ModelWithSequentialFusion,
-    ModelForLinearBNFusion,
-    ManualLinearQATModel,
-    ManualConvLinearQATModel,
     ModelWithFunctionals,
     ModelMultipleOps,
     ModelMultipleOpsNoAvgPool,
@@ -47,11 +35,9 @@ from torch.testing._internal.common_quantization import (
     ResNetBase,
     RNNDynamicModel,
     RNNCellDynamicModel,
-    ModelForFusionWithBias,
     ActivationsTestModel,
     NormalizationTestModel,
     test_only_eval_fn,
-    test_only_train_fn,
     prepare_dynamic,
     convert_dynamic,
     skipIfNoFBGEMM,
@@ -74,8 +60,6 @@ from torch.testing._internal.common_quantized import (
     supported_qengines,
     override_qengines,
 )
-from torch.testing._internal.common_utils import TemporaryFileName
-from torch.testing._internal.common_utils import suppress_warnings
 from torch.testing._internal.jit_utils import JitTestCase
 from hypothesis import given
 from hypothesis import strategies as st
@@ -84,7 +68,6 @@ hu.assert_deadline_disabled()
 
 # Standard library
 from typing import Tuple
-import copy
 import io
 import unittest
 import numpy as np
@@ -1150,208 +1133,7 @@ class TestPostTrainingDynamic(QuantizationTestCase):
             convert_dynamic(model)
             checkHooksIsPresent(model)
 
-
-
-class TestQuantizationAwareTraining(QuantizationTestCase):
-    def test_manual(self):
-        for qengine in supported_qengines:
-            with override_quantized_engine(qengine):
-                model = ManualLinearQATModel(qengine)
-                model = prepare_qat(model)
-                self.checkObservers(model)
-                test_only_train_fn(model, self.train_data)
-                model = convert(model)
-
-                def checkQuantized(model):
-                    self.assertEqual(type(model.fc1), nnq.Linear)
-                    self.assertEqual(type(model.fc2), nnq.Linear)
-                    test_only_eval_fn(model, self.calib_data)
-                    self.checkScriptable(model, self.calib_data)
-                    self.checkNoQconfig(model)
-
-                checkQuantized(model)
-
-                model = quantize_qat(ManualLinearQATModel(qengine), test_only_train_fn,
-                                     [self.train_data])
-                checkQuantized(model)
-
-    def test_eval_only_fake_quant(self):
-        r"""Using FakeQuant in evaluation only mode,
-        this is useful for estimating accuracy loss when we quantize the
-        network
-        """
-        for qengine in supported_qengines:
-            with override_quantized_engine(qengine):
-                model = ManualLinearQATModel(qengine)
-
-                model = prepare_qat(model)
-                self.checkObservers(model)
-
-                model.eval()
-                test_only_eval_fn(model, self.calib_data)
-
-    def test_conv_linear(self):
-        for qengine in supported_qengines:
-            with override_quantized_engine(qengine):
-                model = ManualConvLinearQATModel()
-
-                model = prepare_qat(model)
-                self.checkObservers(model)
-
-                test_only_train_fn(model, self.img_data_2d_train)
-                model = convert(model)
-
-                def checkQuantized(model):
-                    self.assertEqual(type(model.conv), nnq.Conv2d)
-                    self.assertEqual(type(model.fc1), nnq.Linear)
-                    self.assertEqual(type(model.fc2), nnq.Linear)
-                    test_only_eval_fn(model, self.img_data_2d)
-                    self.checkScriptable(model, self.img_data_2d)
-                    self.checkNoQconfig(model)
-
-                checkQuantized(model)
-
-                model = ManualConvLinearQATModel()
-                model = quantize_qat(model, test_only_train_fn, [self.img_data_2d_train])
-                checkQuantized(model)
-
-    def test_train_save_load_eval(self):
-        r"""Test QAT flow of creating a model, doing QAT and saving the quantized state_dict
-        During eval, we first call prepare_qat and conver on the model and then load the state_dict
-        and compare results against original model
-        """
-        for qengine in supported_qengines:
-            with override_quantized_engine(qengine):
-                model = TwoLayerLinearModel()
-                model = torch.quantization.QuantWrapper(model)
-                model.qconfig = torch.quantization.get_default_qat_qconfig(qengine)
-                model = prepare_qat(model)
-
-                fq_state_dict = model.state_dict()
-
-                test_only_train_fn(model, self.train_data)
-                model = convert(model)
-
-                quant_state_dict = model.state_dict()
-
-                x = torch.rand(2, 5, dtype=torch.float)
-                ref = model(x)
-
-                # Create model again for eval. Check result using quantized state_dict
-                model = TwoLayerLinearModel()
-                model = torch.quantization.QuantWrapper(model)
-                model.qconfig = torch.quantization.get_default_qat_qconfig(qengine)
-                torch.quantization.prepare_qat(model, inplace=True)
-                new_state_dict = model.state_dict()
-
-                # Check to make sure the model after prepare_qat has the same state_dict as original.
-                self.assertEqual(set(fq_state_dict.keys()), set(new_state_dict.keys()))
-
-                torch.quantization.convert(model, inplace=True)
-                model.eval()
-                model.load_state_dict(quant_state_dict)
-                out = model(x)
-                self.assertEqual(ref, out)
-
-                # Check model created using prepare has same state dict as quantized state_dict
-                model = TwoLayerLinearModel()
-                model.eval()
-                model = torch.quantization.QuantWrapper(model)
-                model.qconfig = torch.quantization.get_default_qconfig(qengine)
-                torch.quantization.prepare(model, inplace=True)
-                torch.quantization.convert(model, inplace=True)
-                self.assertEqual(set(model.state_dict().keys()), set(quant_state_dict.keys()))
-                model.eval()
-                model.load_state_dict(quant_state_dict)
-                out = model(x)
-                self.assertEqual(ref, out)
-
-    @override_qengines
-    def test_forward_hooks_preserved(self):
-        r"""Test QAT on preserving pre forward and post forward hooks of original model
-        """
-        qengine = torch.backends.quantized.engine
-        model = QuantStubModel()
-        counter = {
-            'pre_forwards': 0,
-            'forwards': 0,
-        }
-
-        def fw_pre_hook(h_module, input):
-            counter['pre_forwards'] += 1
-
-        def fw_hook(h_module, input, output):
-            counter['forwards'] += 1
-
-        model.fc.register_forward_pre_hook(fw_pre_hook)
-        model.fc.register_forward_hook(fw_hook)
-
-        model.qconfig = torch.quantization.get_default_qat_qconfig(qengine)
-        model = prepare_qat(model)
-
-        def checkHooksIsPresent(model, before_convert=True):
-            forward_hooks = 1
-            if before_convert:
-                self.assertEqual(len(model.quant._forward_hooks.values()), 1,
-                                 "Quantization observer hook has disappeared")
-                forward_hooks = 2
-            self.assertObjectIn(fw_pre_hook, model.fc._forward_pre_hooks.values())
-            self.assertObjectIn(fw_hook, model.fc._forward_hooks.values())
-            self.assertEqual(len(model.fc._forward_pre_hooks.values()), 1,
-                             "Extra pre forward hooks have appeared on a layer")
-            self.assertEqual(len(model.fc._forward_hooks.values()), forward_hooks,
-                             "Extra post forward hooks have appeared on a layer")
-
-        checkHooksIsPresent(model, True)
-        x = torch.rand(2, 5, dtype=torch.float)
-        model(x)
-        torch.quantization.convert(model, inplace=True)
-        checkHooksIsPresent(model, False)
-
-    def test_add_scalar_uses_input_qparams(self):
-        class M(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.quant = torch.quantization.QuantStub()
-                self.ff = torch.nn.quantized.FloatFunctional()
-
-            def forward(self, x):
-                x = self.quant(x)
-                x = self.ff.add_scalar(x, 1.0)
-                return x
-
-        m = M()
-        m.qconfig = torch.quantization.default_qconfig
-        mp = torch.quantization.prepare_qat(m)
-        mp(torch.randn(4, 4))
-        mq = torch.quantization.convert(mp)
-        res = mq(torch.randn(4, 4))
-        eps = 1e-5
-        self.assertTrue(torch.abs(mq.quant.scale - res.q_scale()) < eps)
-
-    def test_mul_scalar_uses_input_qparams(self):
-        class M(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.quant = torch.quantization.QuantStub()
-                self.ff = torch.nn.quantized.FloatFunctional()
-
-            def forward(self, x):
-                x = self.quant(x)
-                x = self.ff.mul_scalar(x, 2.0)
-                return x
-
-        m = M()
-        m.qconfig = torch.quantization.default_qconfig
-        mp = torch.quantization.prepare_qat(m)
-        mp(torch.randn(4, 4))
-        mq = torch.quantization.convert(mp)
-        res = mq(torch.randn(4, 4))
-        eps = 1e-5
-        self.assertTrue(torch.abs(mq.quant.scale * 2 - res.q_scale()) < eps)
-
-
-class TestEagerModeOps(QuantizationTestCase):
+class TestEagerModeActivationOps(QuantizationTestCase):
     def _test_activation_op_impl(
             self, float_module_class, quantized_module_class, extra_module_kwargs):
         """ Implementation for testing common activation ops like leaky relu
@@ -1384,102 +1166,6 @@ class TestEagerModeOps(QuantizationTestCase):
     def test_relu(self):
         self._test_activation_op_impl(nn.ReLU, nn.ReLU, {'inplace': False})
 
-
-class TestEagerModeQATOps(QuantizationTestCase):
-    def _test_activation_convert_numerics_impl(self, Act, data):
-        class M(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.act = Act()
-                self.quant = QuantStub()
-                self.dequant = DeQuantStub()
-
-            def forward(self, x):
-                x = self.quant(x)
-                x = self.act(x)
-                x = self.dequant(x)
-                return x
-
-        m = M().train()
-        m.qconfig = default_qat_qconfig
-        m = prepare_qat(m)
-        before_convert = m(data)
-        m = convert(m)
-        after_convert = m(data)
-        self.assertEqual(before_convert, after_convert)
-
-    def test_fixed_qparam_ops(self):
-        class M(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.sigmoid = torch.nn.Sigmoid()
-                self.hardsigmoid = torch.nn.Hardsigmoid()
-                self.tanh = torch.nn.Tanh()
-                self.quant = QuantStub()
-                self.dequant = DeQuantStub()
-
-            def forward(self, x):
-                x = self.quant(x)
-                x = self.sigmoid(x)
-                x = self.hardsigmoid(x)
-                x = self.tanh(x)
-                x = self.dequant(x)
-                return x
-
-        m = M().train()
-        m.qconfig = default_qat_qconfig
-        m = prepare_qat(m)
-        for attr in ['sigmoid', 'hardsigmoid', 'tanh']:
-            self.assertEqual(type(getattr(m, attr).activation_post_process), FixedQParamsFakeQuantize)
-        data = torch.randn(1, 3, 2, 4)
-        before_convert = m(data)
-        m = convert(m)
-        after_convert = m(data)
-        self.assertEqual(before_convert, after_convert)
-        # make sure activation post process is removed
-        for attr in ['sigmoid', 'hardsigmoid', 'tanh']:
-            # verify fake quant module is removd
-            self.assertFalse(hasattr(getattr(m, attr), 'activation_post_process'))
-            # verify that hooks are removed
-            self.assertTrue(len(getattr(m, attr)._forward_hooks.items()) == 0)
-
-        # make sure no fake quantize module is inserted for eval mode
-
-        def checkNoFQModule(m):
-            for attr in ['sigmoid', 'hardsigmoid', 'tanh']:
-                self.assertFalse(hasattr(getattr(m, attr), "activation_post_process"))
-                self.assertTrue(len(getattr(m, attr)._forward_hooks.items()) == 0)
-
-        m = M().eval()
-        m.qconfig = default_qconfig
-        m = prepare(m)
-        checkNoFQModule(m)
-        m = convert(m)
-        checkNoFQModule(m)
-
-    def test_leaky_relu(self):
-        data = torch.randn(1, 3, 2, 4)
-        self._test_activation_convert_numerics_impl(nn.LeakyReLU, data)
-
-    def test_relu(self):
-        class M(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.relu = nn.ReLU()
-
-            def forward(self, x):
-                x = self.relu(x)
-                return x
-
-        m = M().train()
-        m.qconfig = default_qconfig
-        m = prepare_qat(m)
-        # make sure no activation_post_process is inserted for relu
-        self.assertFalse(hasattr(m, "activation_post_process"))
-        m = convert(m)
-        # make sure ReLU module is not changed
-        self.assertTrue(type(m.relu), nn.ReLU)
-
 class TestFunctionalModule(QuantizationTestCase):
     # Histogram Observers are slow, so have no-deadline to ensure test doesn't time out
     @given(train_mode=st.booleans())
@@ -1510,332 +1196,6 @@ class TestFunctionalModule(QuantizationTestCase):
 
         checkQuantized(model)
         self.checkScriptable(model, [[xq]], check_save_load=True)
-
-@skipIfNoFBGEMM
-class TestFusion(QuantizationTestCase):
-    def test_fuse_module_train(self):
-        model = ModelForFusion(default_qat_qconfig).train()
-        # Test step by step fusion
-        model = fuse_modules(model, ['conv1', 'bn1', 'relu1'])
-        model = fuse_modules(model, ['sub1.conv', 'sub1.bn'])
-        self.assertEqual(type(model.conv1), nni.ConvBnReLU2d,
-                         msg="Fused Conv + BN + Relu first layer")
-        self.assertEqual(type(model.bn1), torch.nn.Identity,
-                         msg="Fused Conv + BN + Relu (skipped BN)")
-        self.assertEqual(type(model.relu1), torch.nn.Identity,
-                         msg="Fused Conv + BN + Relu (skipped Relu)")
-
-        self.assertEqual(type(model.sub1.conv), nni.ConvBn2d,
-                         msg="Fused submodule Conv + BN")
-        self.assertEqual(type(model.sub1.bn), torch.nn.Identity,
-                         msg="Fused submodule Conv + BN (skipped BN)")
-        self.assertEqual(type(model.sub2.conv), torch.nn.Conv2d,
-                         msg="Non-fused submodule Conv")
-        self.assertEqual(type(model.sub2.relu), torch.nn.ReLU,
-                         msg="Non-fused submodule ReLU")
-        model = prepare_qat(model)
-        self.checkObservers(model)
-
-        def checkQAT(model):
-            self.assertEqual(type(model.conv1), nniqat.ConvBnReLU2d)
-            self.assertEqual(type(model.bn1), nn.Identity)
-            self.assertEqual(type(model.relu1), nn.Identity)
-            self.assertEqual(type(model.sub1.conv), nniqat.ConvBn2d)
-            self.assertEqual(type(model.sub1.bn), nn.Identity)
-            self.assertEqual(type(model.sub2.conv), nn.Conv2d)
-            self.assertEqual(type(model.sub2.relu), nn.ReLU)
-
-        checkQAT(model)
-        test_only_train_fn(model, self.img_data_1d_train)
-        model = convert(model)
-
-        def checkQuantized(model):
-            self.assertEqual(type(model.conv1), nniq.ConvReLU2d)
-            self.assertEqual(type(model.bn1), nn.Identity)
-            self.assertEqual(type(model.relu1), nn.Identity)
-            self.assertEqual(type(model.sub1.conv), nnq.Conv2d)
-            self.assertEqual(type(model.sub1.bn), nn.Identity)
-            self.assertEqual(type(model.sub2.conv), nn.Conv2d)
-            self.assertEqual(type(model.sub2.relu), nn.ReLU)
-            test_only_eval_fn(model, self.img_data_1d)
-            self.checkNoQconfig(model)
-
-        with self.assertRaisesRegex(RuntimeError, "Could not run 'aten::native_batch_norm' with arguments from the 'QuantizedCPU'"):
-            checkQuantized(model)
-
-        model = ModelForFusion(default_qat_qconfig).train()
-        model = fuse_modules(model, [['conv1', 'bn1', 'relu1'],
-                             ['sub1.conv', 'sub1.bn']])
-        model = quantize_qat(model, test_only_train_fn, [self.img_data_1d_train])
-        with self.assertRaisesRegex(RuntimeError, "Could not run 'aten::native_batch_norm' with arguments from the 'QuantizedCPU'"):
-            checkQuantized(model)
-
-
-    def test_fuse_module_eval(self):
-        model = ModelForFusion(default_qconfig)
-        model.eval()
-        model = fuse_modules(model, [['conv3', 'bn3', 'relu4'],
-                             ['conv1', 'bn1', 'relu1'],
-                             ['conv2', 'relu2'],
-                             ['bn2', 'relu3'],
-                             ['sub1.conv', 'sub1.bn']])
-        self.assertEqual(type(model.conv1), nni.ConvReLU2d,
-                         msg="Fused Conv + BN + Relu first layer (BN is folded)")
-        self.assertEqual(type(model.conv1[0]), nn.Conv2d,
-                         msg="Fused Conv + BN + Relu (Conv + folded BN only)")
-        self.assertEqual(type(model.conv1[1]), nn.ReLU,
-                         msg="Fused Conv + BN + Relu second layer (Relu only)")
-        self.assertEqual(type(model.bn1), nn.Identity,
-                         msg="Fused Conv + BN + Relu second layer (Skipped BN)")
-        self.assertEqual(type(model.relu1), nn.Identity,
-                         msg="Fused Conv + BN + Relu second layer (Skipped Relu)")
-        self.assertEqual(type(model.conv2), nni.ConvReLU3d,
-                         msg="Fused Conv + BN + Relu first layer (BN is folded)")
-        self.assertEqual(type(model.bn2), nni.BNReLU3d,
-                         msg="Fused BN + Relu first layer (Relu is folded))")
-        self.assertEqual(type(model.relu3), nn.Identity,
-                         msg="Fused BN + Relu second layer (Skipped Relu)")
-        self.assertEqual(type(model.conv2[0]), nn.Conv3d,
-                         msg="Fused Conv + BN + Relu (Conv + folded BN only)")
-        self.assertEqual(type(model.conv2[1]), nn.ReLU,
-                         msg="Fused Conv + BN + Relu second layer (Relu only)")
-        self.assertEqual(type(model.relu2), nn.Identity,
-                         msg="Fused Conv + BN + Relu second layer (Skipped Relu)")
-
-        self.assertEqual(type(model.conv3), nni.ConvReLU1d,
-                         msg="Fused Conv + Relu for Conv1d (folded BN)")
-        self.assertEqual(type(model.conv3[0]), nn.Conv1d,
-                         msg="Fused Conv + Relu for Conv1d ")
-        self.assertEqual(type(model.conv3[1]), nn.ReLU,
-                         msg="Fused Conv + Relu for Conv1d")
-        self.assertEqual(type(model.bn3), nn.Identity,
-                         msg="Fused Conv + BN + Relu for Conv1d (Skipped BN)")
-
-        self.assertEqual(type(model.sub1.conv), nn.Conv2d,
-                         msg="Fused submodule Conv + folded BN")
-        self.assertEqual(type(model.sub1.bn), nn.Identity,
-                         msg="Fused submodule (skipped BN)")
-        self.assertEqual(type(model.sub2.conv), nn.Conv2d,
-                         msg="Non-fused submodule Conv")
-        self.assertEqual(type(model.sub2.relu), torch.nn.ReLU,
-                         msg="Non-fused submodule ReLU")
-
-        model = prepare(model)
-        self.checkObservers(model)
-        test_only_eval_fn(model, self.img_data_1d)
-        model = convert(model)
-
-        def checkQuantized(model):
-            self.assertEqual(type(model.conv3), nniq.ConvReLU1d)
-            self.assertEqual(type(model.conv1), nniq.ConvReLU2d)
-            self.assertEqual(type(model.bn1), nn.Identity)
-            self.assertEqual(type(model.relu1), nn.Identity)
-            self.assertEqual(type(model.sub1.conv), nnq.Conv2d)
-            self.assertEqual(type(model.sub1.bn), nn.Identity)
-            self.assertEqual(type(model.sub2.conv), nn.Conv2d)
-            self.assertEqual(type(model.sub2.relu), nn.ReLU)
-            self.assertEqual(type(model.bn2), nniq.BNReLU3d)
-            test_only_eval_fn(model, self.img_data_1d)
-            self.checkNoQconfig(model)
-
-        checkQuantized(model)
-
-        model = ModelForFusion(default_qconfig).eval()
-        model = fuse_modules(model, [['conv1', 'bn1', 'relu1'],
-                             ['conv2', 'relu2'],
-                             ['bn2', 'relu3'],
-                             ['sub1.conv', 'sub1.bn'],
-                             ['conv3', 'bn3', 'relu4']])
-        model = quantize(model, test_only_eval_fn, [self.img_data_1d])
-        checkQuantized(model)
-
-    def test_fusion_sequential_model_train(self):
-        for qengine in supported_qengines:
-            with override_quantized_engine(qengine):
-                model = ModelWithSequentialFusion().train()
-                model.to(torch.float)
-                fuse_modules(model, [['conv1', 'relu1'] ,
-                                     ['features.0.0', 'features.0.1', 'features.0.2'],
-                                     ['features.1.0', 'features.1.1', 'features.1.2'],
-                                     ['features.2.0', 'features.2.1', 'features.2.2'],
-                                     ['classifier.0', 'classifier.1']], inplace=True)
-                self.assertEqual(type(model.conv1), nni.ConvReLU2d,
-                                 msg="Fused Conv + Relu: nni.ConvReLU2d")
-                self.assertEqual(type(model.conv1[0]), nn.Conv2d,
-                                 msg="Fused Conv + Relu: Conv2d")
-                self.assertEqual(type(model.conv1[1]), nn.ReLU,
-                                 msg="Fused Conv + Relu: Relu")
-                self.assertEqual(type(model.relu1), nn.Identity,
-                                 msg="Fused Conv + Relu: Identity")
-                for i in range(3):
-                    self.assertEqual(type(model.features[i][0]), nni.ConvBnReLU2d,
-                                     msg="Fused submodule Conv + folded BN")
-                    self.assertEqual(type(model.features[i][1]), nn.Identity,
-                                     msg="Fused submodule (skipped BN)")
-                    self.assertEqual(type(model.features[i][2]), nn.Identity,
-                                     msg="Non-fused submodule Conv")
-                self.assertEqual(type(model.classifier[0]), nni.LinearReLU)
-                self.assertEqual(type(model.classifier[1]), nn.Identity)
-                model.qconfig = torch.quantization.get_default_qat_qconfig(qengine)
-                prepare_qat(model, inplace=True)
-                self.checkObservers(model)
-                model(self.img_data_2d[0][0])
-
-
-                def checkQAT(model):
-                    self.assertEqual(type(model.conv1), nniqat.ConvReLU2d)
-                    self.assertEqual(type(model.relu1), nn.Identity)
-                for i in range(3):
-                    self.assertEqual(type(model.features[i][0]), nniqat.ConvBnReLU2d,
-                                     msg="Fused submodule Conv + folded BN")
-                    self.assertEqual(type(model.features[i][1]), nn.Identity,
-                                     msg="Fused submodule (skipped BN)")
-                    self.assertEqual(type(model.features[i][2]), nn.Identity,
-                                     msg="Non-fused submodule Conv")
-                self.assertEqual(type(model.classifier[0]), nniqat.LinearReLU)
-                self.assertEqual(type(model.classifier[1]), nn.Identity)
-
-                checkQAT(model)
-                model(self.img_data_2d[1][0])
-                convert(model, inplace=True)
-                model(self.img_data_2d[1][0])
-                self.checkModelWithSequentialQuantized(model)
-
-    def test_fusion_sequential_model_eval(self):
-        for qengine in supported_qengines:
-            with override_quantized_engine(qengine):
-                model = ModelWithSequentialFusion().eval()
-                model.to(torch.float)
-                fuse_modules(model, [['conv1', 'relu1'] ,
-                                     ['features.0.0', 'features.0.1', 'features.0.2'],
-                                     ['features.1.0', 'features.1.1', 'features.1.2'],
-                                     ['features.2.0', 'features.2.1', 'features.2.2'],
-                                     ['classifier.0', 'classifier.1']], inplace=True)
-                self.assertEqual(type(model.conv1), nni.ConvReLU2d,
-                                 msg="Fused Conv + Relu: nni.ConvReLU2d")
-                self.assertEqual(type(model.conv1[0]), nn.Conv2d,
-                                 msg="Fused Conv + Relu: Conv2d")
-                self.assertEqual(type(model.conv1[1]), nn.ReLU,
-                                 msg="Fused Conv + Relu: Relu")
-                self.assertEqual(type(model.relu1), nn.Identity,
-                                 msg="Fused Conv + Relu: Identity")
-                for i in range(3):
-                    self.assertEqual(type(model.features[i][0]), nni.ConvReLU2d,
-                                     msg="Fused submodule Conv + folded BN")
-                    self.assertEqual(type(model.features[i][1]), nn.Identity,
-                                     msg="Fused submodule (skipped BN)")
-                    self.assertEqual(type(model.features[i][2]), nn.Identity,
-                                     msg="Non-fused submodule Conv")
-                self.assertEqual(type(model.classifier[0]), nni.LinearReLU)
-                self.assertEqual(type(model.classifier[1]), nn.Identity)
-                model.qconfig = torch.quantization.get_default_qconfig(qengine)
-                prepare(model, inplace=True)
-                self.checkObservers(model)
-                model(self.img_data_2d[0][0])
-                convert(model, inplace=True)
-                model(self.img_data_2d[1][0])
-                self.checkModelWithSequentialQuantized(model)
-
-    def checkModelWithSequentialQuantized(self, model):
-        self.assertEqual(type(model.conv1), nniq.ConvReLU2d)
-        self.assertEqual(type(model.relu1), nn.Identity)
-        for i in range(3):
-            self.assertEqual(type(model.features[i][0]), nniq.ConvReLU2d)
-            self.assertEqual(type(model.features[i][1]), nn.Identity)
-            self.assertEqual(type(model.features[i][2]), nn.Identity)
-        self.assertEqual(type(model.classifier[0]), nniq.LinearReLU)
-        self.assertEqual(type(model.classifier[1]), nn.Identity)
-
-    def test_fusion_conv_with_bias(self):
-        for qengine in supported_qengines:
-            with override_quantized_engine(qengine):
-                model = ModelForFusionWithBias().train()
-                # output with no fusion.
-                out_ref = model(self.img_data_2d[0][0])
-
-                model.qconfig = QConfig(activation=torch.nn.Identity,
-                                        weight=torch.nn.Identity)
-                model = fuse_modules(model, [["conv1", "bn1", "relu1"],
-                                             ["conv2", "bn2"]])
-                prep_model = prepare_qat(model, inplace=False)
-                # output with fusion but no observers.
-                out_fused = prep_model(self.img_data_2d[0][0])
-                self.assertEqual(out_ref, out_fused)
-
-                model.qconfig = torch.quantization.get_default_qconfig(qengine)
-                prepare_qat(model, inplace=True)
-
-                model(self.img_data_2d[0][0])
-
-                def checkQAT(model):
-                    self.assertEqual(type(model.conv1), nniqat.ConvBnReLU2d)
-                    self.assertEqual(type(model.bn1), nn.Identity)
-                    self.assertEqual(type(model.relu1), nn.Identity)
-                    self.assertEqual(type(model.conv2), nniqat.ConvBn2d)
-                    self.assertEqual(type(model.bn2), nn.Identity)
-
-                checkQAT(model)
-
-
-    def test_fusion_linear_bn_eval(self):
-        model = ModelForLinearBNFusion().train()
-        inp1 = torch.randn(8, 20)
-        inp2 = torch.randn(8, 20)
-
-        # Get some interesting values into the running mean and variance.
-        model(inp1)
-        model.eval()
-        golden = model(inp2)
-
-        model = fuse_modules(model, [["fc", "bn"]])
-        self.assertEqual(type(model.bn), nn.Identity)
-        self.assertEqual(golden, model(inp2))
-
-    def test_forward_hooks_preserved(self):
-        r"""Test case that checks whether forward pre hooks of the first module and
-        post forward hooks of the last module in modules list passed to fusion function preserved.
-        (e.g. before fusion: [nn.Conv2d (with pre forward hooks), nn.BatchNorm2d, nn.ReLU (with post forward hooks)]
-        after fusion: [nni.ConvBnReLU2d (with pre and post hooks), nn.Identity, nn.Identity])
-        """
-        model = ModelForFusion(default_qat_qconfig).train()
-
-        counter = {
-            'pre_forwards': 0,
-            'forwards': 0,
-        }
-        fused = False
-
-        def fw_pre_hook(fused_module_class, h_module, input):
-            if fused:
-                self.assertEqual(type(h_module), fused_module_class,
-                                 "After fusion owner of the first module's forward pre hook is not a fused module")
-            counter['pre_forwards'] += 1
-
-        def fw_hook(fused_module_class, h_module, input, output):
-            if fused:
-                self.assertEqual(type(h_module), fused_module_class,
-                                 "After fusion owner of the last module's forward hook is not a fused module")
-            counter['forwards'] += 1
-
-        # Registering two pre and two post forward hooks, thus expecting counter increment by two each inference
-        model.conv1.register_forward_pre_hook(lambda *args: fw_pre_hook(nni.ConvBnReLU2d, *args))
-        model.sub1.conv.register_forward_pre_hook(lambda *args: fw_pre_hook(nni.ConvBn2d, *args))
-        model.relu1.register_forward_hook(lambda *args: fw_hook(nni.ConvBnReLU2d, *args))
-        model.sub1.bn.register_forward_hook(lambda *args: fw_hook(nni.ConvBn2d, *args))
-
-        test_only_eval_fn(model, self.img_data_1d)
-        self.assertEqual(counter['pre_forwards'], 2 * len(self.img_data_1d))
-        self.assertEqual(counter['forwards'], 2 * len(self.img_data_1d))
-
-        model = fuse_modules(model, ['conv1', 'bn1', 'relu1'])
-        model = fuse_modules(model, ['sub1.conv', 'sub1.bn'])
-
-        fused = True
-        before_fusion_pre_count = counter['pre_forwards']
-        before_fusion_post_count = counter['forwards']
-        test_only_eval_fn(model, self.img_data_1d)
-        self.assertEqual(counter['pre_forwards'] - before_fusion_pre_count, 2 * len(self.img_data_1d))
-        self.assertEqual(counter['forwards'] - before_fusion_post_count, 2 * len(self.img_data_1d))
 
 class TestModelNumerics(QuantizationTestCase):
     def test_float_quant_compare_per_tensor(self):
@@ -1988,343 +1348,6 @@ class TestQuantizeONNXExport(JitTestCase):
         data = torch.from_numpy(data_numpy).to(dtype=torch.float)
         self._test_lower_graph_impl(model, data)
 
-
-class TestDeprecatedJitQuantized(JitTestCase):
-    @skipIfNoFBGEMM
-    def test_rnn_cell_quantized(self):
-        d_in, d_hid = 2, 2
-
-        for cell in [
-            torch.nn.LSTMCell(d_in, d_hid).float(),
-            torch.nn.GRUCell(d_in, d_hid).float(),
-            torch.nn.RNNCell(d_in, d_hid).float(),
-        ]:
-            if isinstance(cell, torch.nn.LSTMCell):
-                num_chunks = 4
-            elif isinstance(cell, torch.nn.GRUCell):
-                num_chunks = 3
-            elif isinstance(cell, torch.nn.RNNCell):
-                num_chunks = 1
-
-            # Replace parameter values s.t. the range of values is exactly
-            # 255, thus we will have 0 quantization error in the quantized
-            # GEMM call. This i s for testing purposes.
-            #
-            # Note that the current implementation does not support
-            # accumulation values outside of the range representable by a
-            # 16 bit integer, instead resulting in a saturated value. We
-            # must take care that in our test we do not end up with a dot
-            # product that overflows the int16 range, e.g.
-            # (255*127+255*127) = 64770. So, we hardcode the test values
-            # here and ensure a mix of signedness.
-            vals = [[100, -155],
-                    [100, -155],
-                    [-155, 100],
-                    [-155, 100],
-                    [100, -155],
-                    [-155, 100],
-                    [-155, 100],
-                    [100, -155]]
-            vals = vals[:d_hid * num_chunks]
-            cell.weight_ih = torch.nn.Parameter(
-                torch.tensor(vals, dtype=torch.float),
-                requires_grad=False)
-            cell.weight_hh = torch.nn.Parameter(
-                torch.tensor(vals, dtype=torch.float),
-                requires_grad=False)
-
-            ref = copy.deepcopy(cell)
-
-            cell = torch.jit.quantized.quantize_rnn_cell_modules(cell)
-            x = torch.tensor([[100, -155],
-                              [-155, 100],
-                              [100, -155]], dtype=torch.float)
-            h0_vals = [[-155, 100],
-                       [-155, 155],
-                       [100, -155]]
-            hx = torch.tensor(h0_vals, dtype=torch.float)
-            if isinstance(cell, torch.jit.quantized.QuantizedLSTMCell):
-                cx = torch.tensor(h0_vals, dtype=torch.float)
-                hiddens = (hx, cx)
-            else:
-                hiddens = hx
-
-            if isinstance(cell, torch.jit.quantized.QuantizedLSTMCell):
-                class ScriptWrapper(torch.jit.ScriptModule):
-                    def __init__(self, cell):
-                        super(ScriptWrapper, self).__init__()
-                        self.cell = cell
-
-                    @torch.jit.script_method
-                    def forward(self, x: torch.Tensor,
-                                hiddens: Tuple[torch.Tensor, torch.Tensor]
-                                ) -> Tuple[torch.Tensor, torch.Tensor]:
-                        return self.cell(x, hiddens)
-            else:
-
-                class ScriptWrapper(torch.jit.ScriptModule):
-                    def __init__(self, cell):
-                        super(ScriptWrapper, self).__init__()
-                        self.cell = cell
-
-                    @torch.jit.script_method
-                    def forward(self, x: torch.Tensor, hiddens: torch.Tensor) -> torch.Tensor:
-                        return self.cell(x, hiddens)
-
-            cell = ScriptWrapper(cell)
-            outs = cell(x, hiddens)
-            cell = self.getExportImportCopyWithPacking(cell)
-
-            outs = cell(x, hiddens)
-            ref_outs = ref(x, hiddens)
-
-            self.assertEqual(len(outs), len(ref_outs))
-            for out, ref_out in zip(outs, ref_outs):
-                torch.testing.assert_allclose(out, ref_out)
-
-    @skipIfNoFBGEMM
-    def test_rnn_quantized(self):
-        d_in, d_hid = 2, 2
-
-        for cell in [
-            torch.nn.LSTM(d_in, d_hid).float(),
-            torch.nn.GRU(d_in, d_hid).float(),
-        ]:
-
-            # Replace parameter values s.t. the range of values is exactly
-            # 255, thus we will have 0 quantization error in the quantized
-            # GEMM call. This i s for testing purposes.
-            #
-            # Note that the current implementation does not support
-            # accumulation values outside of the range representable by a
-            # 16 bit integer, instead resulting in a saturated value. We
-            # must take care that in our test we do not end up with a dot
-            # product that overflows the int16 range, e.g.
-            # (255*127+255*127) = 64770. So, we hardcode the test values
-            # here and ensure a mix of signedness.
-            vals = [[100, -155],
-                    [100, -155],
-                    [-155, 100],
-                    [-155, 100],
-                    [100, -155],
-                    [-155, 100],
-                    [-155, 100],
-                    [100, -155]]
-            if isinstance(cell, torch.nn.LSTM):
-                num_chunks = 4
-            elif isinstance(cell, torch.nn.GRU):
-                num_chunks = 3
-            vals = vals[:d_hid * num_chunks]
-            cell.weight_ih_l0 = torch.nn.Parameter(
-                torch.tensor(vals, dtype=torch.float),
-                requires_grad=False)
-            cell.weight_hh_l0 = torch.nn.Parameter(
-                torch.tensor(vals, dtype=torch.float),
-                requires_grad=False)
-
-            ref = copy.deepcopy(cell)
-            cell_int8 = torch.jit.quantized.quantize_rnn_modules(cell, dtype=torch.int8)
-            cell_fp16 = torch.jit.quantized.quantize_rnn_modules(cell, dtype=torch.float16)
-
-            niter = 10
-            x = torch.tensor([[100, -155],
-                              [-155, 100],
-                              [100, -155]], dtype=torch.float).unsqueeze(0).repeat(niter, 1, 1)
-            h0_vals = [[-155, 100],
-                       [-155, 155],
-                       [100, -155]]
-            hx = torch.tensor(h0_vals, dtype=torch.float).unsqueeze(0)
-            cx = torch.tensor(h0_vals, dtype=torch.float).unsqueeze(0)
-
-            if isinstance(ref, torch.nn.LSTM):
-                hiddens = (hx, cx)
-            elif isinstance(ref, torch.nn.GRU):
-                hiddens = hx
-
-            ref_out, ref_hid = ref(x, hiddens)
-
-            # Compare int8 quantized to unquantized
-            output_int8, final_hiddens_int8 = cell_int8(x, hiddens)
-
-            torch.testing.assert_allclose(output_int8, ref_out)
-            for out, ref in zip(final_hiddens_int8, ref_hid):
-                torch.testing.assert_allclose(out, ref)
-
-            # Compare fp16 quantized to unquantized
-            output_fp16, final_hiddens_fp16 = cell_fp16(x, hiddens)
-
-            torch.testing.assert_allclose(output_fp16, ref_out)
-            for out, ref in zip(final_hiddens_fp16, ref_hid):
-                torch.testing.assert_allclose(out, ref)
-
-            def compare_quantized_unquantized(ScriptWrapper, cell):
-                wrapper = ScriptWrapper(cell)
-
-                # Compare quantize scripted module to unquantized
-                script_out, script_hid = wrapper(x, hiddens)
-                torch.testing.assert_allclose(script_out, ref_out)
-                for out, ref in zip(script_hid, ref_hid):
-                    torch.testing.assert_allclose(out, ref)
-
-                # Compare export/import to unquantized
-                export_import_wrapper = self.getExportImportCopyWithPacking(wrapper)
-                ei_out, ei_hid = export_import_wrapper(x, hiddens)
-                torch.testing.assert_allclose(ei_out, ref_out)
-                for out, ref in zip(ei_hid, ref_hid):
-                    torch.testing.assert_allclose(out, ref)
-
-            if isinstance(cell, torch.jit.quantized.QuantizedGRU):
-                class ScriptWrapper(torch.jit.ScriptModule):
-                    def __init__(self, cell):
-                        super(ScriptWrapper, self).__init__()
-                        self.cell = cell
-
-                    @torch.jit.script_method
-                    def forward(self, x: torch.Tensor, hiddens: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
-                        return self.cell(x, hiddens)
-
-                compare_quantized_unquantized(ScriptWrapper, cell)
-            elif isinstance(cell, torch.jit.quantized.QuantizedLSTM):
-                for cell in [cell_int8, cell_fp16]:
-                    class ScriptWrapper(torch.jit.ScriptModule):
-                        def __init__(self, cell):
-                            super(ScriptWrapper, self).__init__()
-                            self.cell = cell
-
-                        @torch.jit.script_method
-                        def forward(self, x, hiddens):
-                            # type: (torch.Tensor, Tuple[torch.Tensor, torch.Tensor])
-                            #        -> Tuple[torch.Tensor, Tuple[torch.Tensor, torch.Tensor]]
-                            return self.cell(x, hiddens)
-                    compare_quantized_unquantized(ScriptWrapper, cell)
-
-    if 'fbgemm' in torch.backends.quantized.supported_engines:
-        # Suppression: using deprecated quant api
-        @suppress_warnings
-        def test_quantization_modules(self):
-            K1, N1 = 2, 2
-
-            class FooBar(torch.nn.Module):
-                def __init__(self):
-                    super(FooBar, self).__init__()
-                    self.linear1 = torch.nn.Linear(K1, N1).float()
-
-                def forward(self, x):
-                    x = self.linear1(x)
-                    return x
-
-            fb = FooBar()
-            fb.linear1.weight = torch.nn.Parameter(
-                torch.tensor([[-150, 100], [100, -150]], dtype=torch.float), requires_grad=False)
-            fb.linear1.bias = torch.nn.Parameter(torch.zeros_like(fb.linear1.bias), requires_grad=False)
-
-            x = (torch.rand(1, K1).float() - 0.5) / 10.0
-            value = torch.tensor([[100, -150]], dtype=torch.float)
-
-            y_ref = fb(value)
-
-            fb_int8 = torch.jit.quantized.quantize_linear_modules(fb)
-            traced_int8 = torch.jit.trace(fb_int8, (x,))
-            fb_int8 = self.getExportImportCopyWithPacking(traced_int8)
-            y_int8 = fb_int8(value)
-
-            fb_fp16 = torch.jit.quantized.quantize_linear_modules(fb, torch.float16)
-            traced_fp16 = torch.jit.trace(fb_fp16, (x,))
-            fb_fp16 = self.getExportImportCopyWithPacking(traced_fp16)
-            y_fp16 = fb_fp16(value)
-
-            torch.testing.assert_allclose(y_int8, y_ref, rtol=0.0001, atol=1e-3)
-            torch.testing.assert_allclose(y_fp16, y_ref, rtol=0.0001, atol=1e-3)
-
-    def _test_pickle_checkpoint_qtensor(self, device):
-        with TemporaryFileName() as fname:
-            class M(torch.jit.ScriptModule):
-                __constants__ = ['fname']
-
-                def __init__(self):
-                    super(M, self).__init__()
-                    self.fname = fname
-
-                @torch.jit.script_method
-                def forward(self, x, y):
-                    torch.save((x, y), self.fname)
-                    return y
-
-            q = torch.quantize_per_tensor(
-                torch.rand(2, 3, dtype=torch.float), scale=0.1, zero_point=10, dtype=torch.quint8).to(device)
-            qc = torch.quantize_per_channel(
-                torch.rand(2, 3, dtype=torch.float),
-                scales=torch.tensor([0.1, 0.5, 0.01]),
-                zero_points=torch.tensor([10, 0, 20]),
-                axis=1, dtype=torch.quint8).to(device)
-            m = M()
-            m(q, qc)
-            with open(fname, "rb") as handle:
-                loaded_q, loaded_qc = torch.load(fname)
-                self.assertEqual(loaded_q, q)
-                self.assertEqual(loaded_qc, qc)
-
-    def test_pickle_checkpoint_qtensor(self):
-        self._test_pickle_checkpoint_qtensor('cpu')
-
-    def test_serialize_qtensor(self):
-        class SimpleQTensor(torch.jit.ScriptModule):
-            def __init__(self, per_channel):
-                super(SimpleQTensor, self).__init__()
-                x = torch.rand(5, 5).float()
-                if not per_channel:
-                    x_q = torch.quantize_per_tensor(x, 0.2, 10, torch.quint8)
-                else:
-                    s = torch.rand(5, dtype=torch.float64) + 0.1
-                    zp = torch.randint(5, 15, (5,))
-                    x_q = torch.quantize_per_channel(x, s, zp, 1, torch.quint8)
-                self.register_buffer('x', x_q)
-
-            @torch.jit.script_method
-            def forward(self):
-                return self.x
-
-        for per_channel in [False, True]:
-            model = SimpleQTensor(per_channel)
-            buffer = io.BytesIO()
-            torch.jit.save(model, buffer)
-            buffer.seek(0)
-            model_loaded = torch.jit.load(buffer)
-            self.assertEqual(model_loaded(), model())
-
-    @skipIfNoFBGEMM
-    def test_erase_class_tensor_shapes(self):
-        class Linear(torch.nn.Module):
-            def __init__(self, in_features, out_features):
-                super(Linear, self).__init__()
-                qweight = torch._empty_affine_quantized(
-                    [out_features, in_features], scale=1, zero_point=0,
-                    dtype=torch.qint8)
-                self._packed_weight = torch.ops.quantized.linear_prepack(qweight)
-
-            @torch.jit.export
-            def __getstate__(self):
-                return (torch.ops.quantized.linear_unpack(self._packed_weight)[0], self.training)
-
-            def forward(self):
-                return self._packed_weight
-
-            @torch.jit.export
-            def __setstate__(self, state):
-                self._packed_weight = torch.ops.quantized.linear_prepack(state[0])
-                self.training = state[1]
-
-            @property
-            def weight(self):
-                return torch.ops.quantized.linear_unpack(self._packed_weight)[0]
-
-            @weight.setter
-            def weight(self, w):
-                self._packed_weight = torch.ops.quantized.linear_prepack(w)
-
-        with torch._jit_internal._disable_emit_hooks():
-            x = torch.jit.script(Linear(10, 10))
-            torch._C._jit_pass_erase_shape_information(x.graph)
 
 if __name__ == '__main__':
     raise RuntimeError("This test file is not meant to be run directly, use:\n\n"
